+++ iMovie/Scenes/Movies/MoviesView.swift
@@ -9,14 +9,6 @@ import SwiftUI
 
 struct MoviesView: View {
 
-    // MARK: - Private Methods
-
-    private func topSafeAreaInsets() -> CGFloat {
-        guard let key = UIApplication.shared.windows.first?.safeAreaInsets.top else { return 5.0 }
-
-        return key + 10.0
-    }
-
     // MARK: - Body
 
     var body: some View {
@@ -30,12 +22,21 @@ struct MoviesView: View {
 
                 Spacer()
 
-                ContentPickerView()
+                CategoryPickerView()
             }
         }
         .ignoresSafeArea(.all, edges: .top)
     }
+
+    // MARK: - Private Methods
+
+    private func topSafeAreaInsets() -> CGFloat {
+        guard let key = UIApplication.shared.windows.first?.safeAreaInsets.top else { return 5.0 }
+
+        return key + 10.0
+    }
 }
+
 struct MovieView_Previews: PreviewProvider {
     static var previews: some View {
         MoviesView()
